hp.map {harrypotter}	R Documentation

Original 'Harry Potter' colour map

Description

A dataset containing some colour palettes inspired on the Harry Potter Universe

Usage

hp.map

Format

A data frame containing all the colours used in the palette:

V1: Red value

V2: Green value

V3: Blue value

option: It is intended to be a general option for choosing the specific colour palette.

[Package harrypotter version 2.1.0 Index]