hp.map {harrypotter}	R Documentation

Original 'Harry Potter' colour map

Description

A dataset containing the colour palettes of each house of Hogwarts school.

Usage

hp.map

Format

A data frame containing all the colours used in the palette:

V1: Red value

V2: Green value

V3: Blue value

house: Refers to the house of Hogwarts. It is intended to be a general option for choosing the specific colour palette.

[Package harrypotter version 2.0.0 Index]